fix failing test and add more tests
