+++ src/button-group/__integ__/button-group.test.ts
@@ -79,6 +79,10 @@ test(
     await expect(page.getText(buttonGroup.findTooltip().toSelector())).resolves.toBe('Liked');
 
     await page.click(createWrapper().find('[data-testid="focus-on-copy"]').toSelector());
+    await expect(page.isFocused(copyButton.toSelector())).resolves.toBe(true);
+
+    await page.keys(['Tab', 'Enter']);
+    await expect(page.isFocused(copyButton.toSelector())).resolves.toBe(true);
     await expect(page.getText(buttonGroup.findTooltip().toSelector())).resolves.toBe('Copy');
   })
 );
@@ -142,5 +146,33 @@ test(
 
     await page.click(sendButton.toSelector());
     await expect(page.getText(buttonGroup.findTooltip().toSelector())).resolves.toBe('Send');
+  })
+);
+
+test(
+  'does not show tooltip over a menu after a menu item is clicked',
+  setup({}, async page => {
+    await page.click(actionsMenu.toSelector());
+    await expect(page.isExisting(actionsMenu.findOpenDropdown().toSelector())).resolves.toBe(true);
+    await expect(page.isExisting(buttonGroup.findTooltip().toSelector())).resolves.toBe(false);
+
+    await page.click(actionsMenu.findItemById('edit').toSelector());
+    await expect(page.isExisting(actionsMenu.findOpenDropdown().toSelector())).resolves.toBe(false);
+    await expect(page.isExisting(buttonGroup.findTooltip().toSelector())).resolves.toBe(false);
+  })
+);
+
+test(
+  'hides tooltip when focus moves to the next component from menu dropdown',
+  setup({}, async page => {
+    await page.click(actionsMenu.toSelector());
+    await expect(page.isExisting(actionsMenu.findOpenDropdown().toSelector())).resolves.toBe(true);
+
+    await page.keys(['ArrowDown', 'ArrowDown']);
+    await expect(page.getFocusedElementText()).resolves.toBe('Edit');
+
+    await page.keys(['Tab']);
+    await expect(page.getFocusedElementText()).resolves.toBe('Focus on copy');
+    await expect(page.isExisting(buttonGroup.findTooltip().toSelector())).resolves.toBe(false);
   })
 );
