+++ src/button-group/item-element.tsx
@@ -3,7 +3,8 @@
 import React, { forwardRef, useEffect, useImperativeHandle, useRef } from 'react';
 
 import { ButtonProps } from '../button/interfaces.js';
-import { ClickDetail, fireCancelableEvent, NonCancelableEventHandler } from '../internal/events/index.js';
+import { ClickDetail, fireCancelableEvent, NonCancelableEventHandler } from '../internal/events';
+import { useFocusVisibleState } from '../internal/hooks/use-focus-visible-state/index.js';
 import IconButtonItem from './icon-button-item.js';
 import { ButtonGroupProps } from './interfaces.js';
 import MenuDropdownItem from './menu-dropdown-item.js';
@@ -29,6 +30,7 @@ const ItemElement = forwardRef(
   ) => {
     const containerRef = useRef<HTMLDivElement>(null);
     const buttonRef = useRef<HTMLButtonElement>(null);
+    const focusVisible = useFocusVisibleState();
 
     useImperativeHandle(ref, () => ({
       focus: () => {
@@ -75,10 +77,9 @@ const ItemElement = forwardRef(
     };
 
     const onShowTooltipHard = (show: boolean) => {
-      if (!show && item.id !== tooltip?.item) {
+      if (show && !focusVisible) {
         return;
       }
-
       setTooltip(show ? { item: item.id, feedback: false } : null);
     };
