+++ src/internal/components/chart-plot/focus-outline.tsx
@@ -1,9 +1,8 @@
 // Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
 // SPDX-License-Identifier: Apache-2.0
-import React, { useEffect, useRef, useState } from 'react';
+import React, { useEffect, useRef } from 'react';
 
-import { isModifierKey } from '@cloudscape-design/component-toolkit/internal';
-
+import { useFocusVisibleState } from '../../hooks/use-focus-visible-state';
 import { Offset } from '../interfaces';
 
 import styles from './styles.css.js';
@@ -12,32 +11,6 @@ export interface FocusOutlineProps {
   elementKey?: null | string | number | boolean;
   elementRef?: React.RefObject<SVGSVGElement | SVGGElement>;
   offset?: Offset;
-}
-
-function useFocusVisibleState() {
-  const [focusVisible, setFocusVisible] = useState(false);
-  useEffect(() => {
-    function handleMousedown() {
-      return setFocusVisible(false);
-    }
-
-    function handleKeydown(event: KeyboardEvent) {
-      // we do not want to highlight focused element
-      // when special keys are pressed
-      if (!isModifierKey(event)) {
-        setFocusVisible(true);
-      }
-    }
-
-    document.addEventListener('mousedown', handleMousedown);
-    document.addEventListener('keydown', handleKeydown);
-    return () => {
-      document.removeEventListener('mousedown', handleMousedown);
-      document.removeEventListener('keydown', handleKeydown);
-    };
-  });
-
-  return focusVisible;
 }
 
 export default function FocusOutline({ elementKey, elementRef, offset = 0 }: FocusOutlineProps) {
